requirements.txt
beautifulsoup4==4.11.1
emoji==2.8.0
fastapi==0.104.1
fasttext==0.9.2
huggingface_hub==0.17.2
keybert==0.7.0
requests==2.31.0
transformers==4.24.0
uvicorn==0.23.2
openai==1.3.9
pymongo==4.6.1
pydantic==2.4.2
pydantic-core==2.10.1
paho-mqtt==1.6.1
redis==4.5.4
langid==1.1.6
python-dotenv==1.0.1
ray==2.9.3
pytest==8.1.1
scipy==1.10.1
gensim==4.3.2
python-keycloak==3.3.0
pydantic-settings==2.2.1